requirements.txt
transformers>=4.25.1
torch>=1.13.1
streamlit>=1.18.0
pdfplumber>=0.7.6
python-docx>=0.8.11
PyMuPDF>=1.21.1
scikit-learn>=1.2.0
textstat>=0.7.3
spacy>=3.4.0
nltk>=3.7